Update mainp1.py
Added functions for my two questions, (privileged users 3 and 4) and also added a check_privileged function  and usertasks function to start of the interface a bit. Please feel free to check over the code to see if i made any syntax errors or did something wrong - I was mostly guessing at the syntax based on the code that was already there, but probably made some errors. In particular, I used a for loop to go through the list of uids and couldn't find much explanation on the syntax when i looked it up so it may be a bit off